feat(thread-projection): complete approval + truncated semantic fields (P1 step 3)
The shared model now carries the remaining wire semantics the audit maps to
P1 (G1/G3/G13/G15 already landed in step 2; this closes approval.deletes,
G2's semantic half, and G13's truncation detection):

- New src/approval.ts: the approval semantic model single-sourced — guarded
  parseAction/parseSaveSuggestion/unwrapNestedJson/formatRequestedAction plus
  approvalPaths/approvalDeletes/approvalCommand. Desktop's approvalPayload.ts
  and the ApprovalAction/ApprovalSaveSuggestion types in storage/types.ts are
  deleted; ApprovalPrompt imports the package. Mobile's inline path extractors
  in TimelineCard are replaced by the same package helpers.
- The live projector detects agent_end reason "incomplete" and surfaces it as
  AssistantRunProjection.truncated / AgentMessage.truncated, so both ends can
  render a truncated reply as failed instead of a clean completion.

Both ends green: desktop tsc/lint clean, 254 vitest tests (3 new truncated
cases); mobile tsc/lint clean, 71 jest tests.
