+++ thread-projection/src/model.ts
@@ -83,6 +83,12 @@ export interface AgentMessage {
   /** The reply was interrupted by the user (its run was cancelled mid-stream). */
   stopped?: boolean;
   /**
+   * The stream ended before the model finished (`agent_end` reason
+   * "incomplete"): the text is a truncated prefix, not a finished answer, and
+   * must not render as a clean completion.
+   */
+  truncated?: boolean;
+  /**
    * The model is mid-reasoning with nothing visible yet. Drives the footer
    * "thinking…" hint (only while streaming and the show-thinking setting is off).
    */
